Make OSAs orphans in Sphinx terms
OSAs are not included in built documentation directly but as a symlinked
documents. This results in warning about not including source of the
symlink in rendered docs: "WARNING: document isn't included in any
toctree".

Making them "orphans" suppresses this warning for both source advisories
as well as symlinks. They are included in table of contents by a
wildcard so in the end none will get lost.

Issue-ID: SECCOM-264
Change-Id: Ia2920c3d289c402c74de01d21818c67424664f14
Signed-off-by: Pawel Wieczorek <p.wieczorek2@samsung.com>